+++ app/models/subject.rb
@@ -35,15 +35,11 @@ class Subject < ApplicationRecord
   has_many :adverse_events, -> { current }
   has_many :ae_adverse_events, -> { current }
   has_many :randomizations, -> { current }
+  has_many :medications, -> { current }
   has_many :sheets, -> { current }
   has_many :subject_events, -> { joins(:event).order(:event_date, "events.position") }
 
   # Methods
-
-  def medications
-    []
-  end
-
   def self.searchable_attributes
     %w(subject_code)
   end